Added checks before writes and membank setting
The related issue is Issue #10 in lis2dux12-pid repository (see
https://github.com/STMicroelectronics/lis2dux12-pid/issues/10).

If a reads fails updates pointers with random data. Checks are added
to avoid such issue.

Another set of analogous checks was added to the membank so it jumps
to the exit condition (resetting the membank to its correct state)
before returning the error.

Signed-off-by: Gioele Fiorenza <gioele.fiorenza2000@gmail.com>